requirements.txt
py-clob-client==0.17.0
python-dotenv==1.0.1
websockets==12.0
httpx==0.27.0
anthropic==0.40.0
pydantic==2.9.0
structlog==24.4.0
apscheduler==3.10.4
numpy==1.26.4
